requirements.txt
# FastAPI and server
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9

# API clients
groq>=0.11.0
voyageai>=0.3.3
qdrant-client>=1.11.0

# Configuration and validation
pydantic>=2.9.0
pydantic-settings>=2.5.0
python-dotenv>=1.0.1

# Async HTTP
httpx>=0.27.0
aiofiles>=24.1.0

# Logging
python-json-logger>=2.0.7

# Document processing (for ingestion)
tiktoken>=0.7.0
